Rename Brush -> BrushX
